Create all components needed for second view.
The second view is the main view that the student is going to use and play.

Task is a unit of work which a student has to complete, there are going to be 3 of the in the tool(easy, medium, hard).
It's the container for the progress bar, beats.

The task has separate components inside it such as the progress bar, the beats, indicators and so on.

Build components for second view such as Lives (contains 3 RadioButtons which are the lives), NextButton, Rails, Tasks.

Also added general RadioButtons component to reuse for the beat buttons in the task and the Lives (they have 3 states, turned off, green or red).
StatusIndicator separate component created.

Add movement to progress bar based on audio current time, the progress bar is closely tied with the AudioManager to manage.

Also added play pause functionality.
Using the audio context to manage sounds.

Added events to beats and ability to adjust lives.
Also deleted .DS_Store files and added it in gitignore.