Fake progress bar, task switching, track switching
If a student fails 3 or more then reset the song and beats.

Fake progress to account for more bars.

We realized looping the same short voice clip over and over is boring, so we decided to use songs with more bars, but only show as much as we can fit. In a 4/4 signature where the full track has 8 bars, the progress bar only shows the first 4 even though all 8 are played and looping.

Add task switching with transitions when the user presses the next button.